Removed "TelosB application"; we made a model, not an application
The CC2420 example in the scratch telosb program uses the CC2420 on off application, not a TelosB specific application. Another example of a model is the sixlowpan model which uses the ping6 application in its example. We don't need an "Application" per se, but rather an example use case, which is provided in scratch/telosb-three-nodes.cc. Alternatively, we can create an example in processing/examples that does the same, but we can't compile examples due to errors caused by the CC2420 module among other things.