Untracked many file, commented out Stein's code in ad-hoc wifi module
Stein's old code there causes circular dependency conflicts with the CC2420 module. The good thing is that that code must not be active when running the TelosB model because it is specific to Stein's and Øystein's models. The bad thing is that it makes things more complex, and from now on, the code will be different when running the old models versus running TelosB.