fixed the ThenOperator, hopefully
Now, the constraints work in ThenOperator. Additionally, 4 event #1 in the sequence and a single event #2 causes 4 complex events to be generated. Therefore, running the random event generation with 100 random events causes in this commit over 600 complex events. 1% of the rules (10) are triggered for each event, which means rules are triggered 1000 times. If only maximum one complex event could be generated for each event #2, the number of total generated complex events would be lower than 600.