created complex events get timestamp and are processed by engine
The latter point means that if the query AthenB results in a complex event C1, C1 gets processed by the CEP engine that created it. This enables us to link together different types of queries and model queries with more than two states. For instance AthenBthenCthenD can be modeled as AthenB=C1, C1thenC=C2, and C2thenD. The limitation to this is that the complex events C1 and C2 must never be generated in another way than through these two queries. If we can do that, we are able to model very complex queries, including a mix of multiple And, Then and Or operators.