first gritty work to chain together queries
the main thing I've done is to prevent processing to occur to an event that has the skipProcessing flag set to true. This flag is set to true if the query that produces it has the isFinalWithinNode flag set to true. If that flag is set to true, it means that the complex event simply represents the result of that query and is not an actual event. The checks to see if the processing of an event should be skipped are very ugly, and I think we can make a much prettier solution.