dequeuing from current queue fixes crash
when in a loop, you apparently can't dequeue from any specific queue, and you might have to dequeue from a queue when in a loop, or else the simulation crashes. I don't quite understand the logic, but it confirms that the code needs refactoring. When we have time, I guess