fixed crashes due to errors in the device file
So an inaccuracy that I probably won't fix is that each thread should handle its own complex events, but I find that hard to do currently. Perhaps each thread should have its own triggers for packets. And somehow, we should be able to diverge paths so that thread 1 and 2 can work on same packet, but use different triggers. That must be left for future work, however.