added options to simulation and added 10 second wait before generating events
the delay is to make sure that the complex queries reach Node C and Node A/B get a proper route to C before they attempt to transmit events to it. The problem that occurs is that those nodes will by default wait 100 ms to transmit those events until the route is set, which causes the processing model to get out of sync, and thus the SEM to transmit packets is executed before the packets are placed in the queue, and no packets are sent.