Now we use ExecutionInfo queues instead of Packet queues, but with issues
Nothing is compiling because of a strange linking error that complains that the GetSize method of ExecutionInfo can't be found or resolved in some way. It has been implemented and I think it should work. I did try to move the executioninfo files to the network module, but it didn't fix it. There should be a clear reason why.