now the simulation is executing without error, but probably not correctly
We still need to tweak how the m_executionInfo object fits in ProgramLocation and how other threads inherit it by creating a completely new ExecutionInfo object based on the one it inherits. Currently, we don't create a new one and just use the one from the queue. We need to create a new ExecutionInfo object every time it is dequeued from a queue.