working some more on the improved trigger system
trying to get all the triggers to work for the new system.
Some issues have surfaced, however, like the unpredictable access to variables like CepEvent and Packet. I think we need to make the modeling language a bit more like a programming language where variables are accessible in a scope, and so on and so forth. Currently, it works almost like that, but not quite.