enabled more logging and fixed device file logic
Before, the curCepEvent variable would be overwritten when HIRQ-1 is invoked, which can overwrite it for the thread running check-constraints. We now dequeue the event from a cep event queue. We don't explicitly dequeue cep events when doing the handle-cepops FSM, but instead dequeue a cep query, which dequeues an ExecutionInfo object which contains the relevant CepEvent.