added Siddhi and T-Rex specific CEPEngine objects
Not entirely sure if we want that because it means a higher modeling effort. The good part is that we will allow for differences in functionality, and it illustrates how we can use different FogServices. I think it's good to have for now, and maybe we will use them later on as the query processing gets complex.